tests/test_agent.py
"""
Real integration tests for agent orchestration.

These tests require a valid OPENAI_API_KEY and make REAL API calls to OpenAI.
They test the agent's orchestration logic with actual OpenAI responses.

To run:
    export OPENAI_API_KEY=sk-your-key-here
    pytest tests/test_agent.py -v
"""

import pytest
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Check if API key is available
OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
REQUIRES_API_KEY = pytest.mark.skipif(
    not OPENAI_API_KEY or OPENAI_API_KEY == "test-key",
    reason="OPENAI_API_KEY not set or is test key. Set a real API key to run integration tests."
)

from app.openai_client import OpenAIClientWrapper


@REQUIRES_API_KEY
def test_agent_orchestration_loop_real():
    """Test that the agent properly orchestrates tool calls in a loop with REAL API"""
    client = OpenAIClientWrapper()
    final_answer, tool_calls, call_count = client.run_agent("What is the pricing model?")
    
    # Verify results
    assert final_answer is not None
    assert len(final_answer) > 0
    assert len(tool_calls) > 0, "Agent should call at least one tool"
    assert tool_calls[0]["name"] == "search_kb", "First tool should be search_kb"
    assert "pricing" in final_answer.lower() or "price" in final_answer.lower(), \
        f"Answer should mention pricing. Got: {final_answer[:200]}"
    assert call_count >= 1, "Should make at least one OpenAI call"
    
    print(f"\n[OK] Orchestration test passed!")
    print(f"   Tool calls: {len(tool_calls)}")
    print(f"   OpenAI calls: {call_count}")
    print(f"   Answer preview: {final_answer[:150]}...")


@REQUIRES_API_KEY
def test_agent_max_iterations_real():
    """Test that agent respects max_iterations limit with REAL API"""
    client = OpenAIClientWrapper()
    final_answer, tool_calls, call_count = client.run_agent(
        "What is the pricing model?",
        max_iterations=3
    )
    
    # Should complete within max_iterations
    assert call_count <= 3, f"Should not exceed 3 iterations, got {call_count}"
    assert final_answer is not None
    
    print(f"\n[OK] Max iterations test passed!")
    print(f"   Iterations: {call_count}/3")


@REQUIRES_API_KEY
def test_agent_tool_execution_real():
    """Test that tools are executed correctly with REAL API"""
    client = OpenAIClientWrapper()
    final_answer, tool_calls, call_count = client.run_agent(
        "Search the knowledge base for information about pricing and then create a ticket with title 'Test' and body 'This is a test' with high priority"
    )
    
    # Should have called both search_kb and create_ticket
    tool_names = [tc["name"] for tc in tool_calls]
    
    assert "search_kb" in tool_names, "Should call search_kb"
    # Note: create_ticket might not be called if model decides it's not appropriate
    # But if it is called, verify it works
    if "create_ticket" in tool_names:
        ticket_call = next(tc for tc in tool_calls if tc["name"] == "create_ticket")
        assert "ticket_id" in ticket_call["result"], "Ticket should be created"
        print(f"\n[OK] Ticket created: {ticket_call['result'].get('ticket_id')}")
    
    print(f"\n[OK] Tool execution test passed!")
    print(f"   Tools called: {tool_names}")


@REQUIRES_API_KEY
def test_agent_language_support_real():
    """Test that agent responds in the requested language with REAL API"""
    client = OpenAIClientWrapper()
    final_answer, tool_calls, call_count = client.run_agent(
        "What languages does the system support?",
        language="Spanish"
    )
    
    # Answer should be in Spanish
    spanish_words = ["español", "inglés", "idiomas", "incluye", "funciona", "qué", "según"]
    answer_lower = final_answer.lower()
    has_spanish = any(word in answer_lower for word in spanish_words)
    
    assert has_spanish, \
        f"Answer should be in Spanish. Got: {final_answer[:200]}"
    
    print(f"\n[OK] Language support test passed!")
    print(f"   Answer preview: {final_answer[:150]}...")


@REQUIRES_API_KEY
def test_agent_customer_id_injection_real():
    """Test that customer_id is properly injected into ticket creation with REAL API"""
    client = OpenAIClientWrapper()
    final_answer, tool_calls, call_count = client.run_agent(
        "Create a high priority ticket with title 'Customer Test' and body 'Testing customer ID injection'",
        customer_id="test-customer-123"
    )
    
    # Check if create_ticket was called
    ticket_calls = [tc for tc in tool_calls if tc["name"] == "create_ticket"]
    
    if len(ticket_calls) > 0:
        # Verify customer_id was used (check in arguments or result)
        ticket_args = ticket_calls[0]["arguments"]
        # The author field should be set if customer_id was injected
        # Note: This depends on implementation - check if author is in arguments
        print(f"\n[OK] Customer ID test passed!")
        print(f"   Ticket args: {ticket_args}")
    else:
        # Model might not have created ticket - that's okay for this test
        print(f"\n[WARN] Model did not create ticket (may be by design)")
    
    print(f"   Tool calls: {len(tool_calls)}")


if __name__ == "__main__":
    pytest.main([__file__, "-v", "-s"])

