requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai==1.12.0
pydantic==2.5.3
python-dotenv==1.0.0
pytest==7.4.4
httpx==0.26.0
requests==2.31.0
slowapi==0.1.9
pymongo==4.6.1
boto3==1.34.0
anthropic==0.34.0
google-generativeai==0.8.0

